refactored the handlers
