+++ templates/signin_signup.html
@@ -68,7 +68,7 @@
                   </div>
                   <div class="pv2 w-100">
                     <label class="dn w3">Password</label>
-                    <input type="text" name="password" class="pa3 w-100 bg-near-white" placeholder="Password"/>
+                    <input type="password" name="password" class="pa3 w-100 bg-near-white" placeholder="Password"/>
                   </div>
                   <div class="tr">
                     <button type="submit" class="pv3 ph4 bg-navy white bw0 shadow-4">Login</button>
@@ -88,7 +88,7 @@
                   </div>
                   <div class="pv2 w-100">
                     <label class="dn w3">Password</label>
-                    <input type="text" name="passwd" class="pa3 w-100 bg-near-white" placeholder="Password"/>
+                    <input type="password" name="passwd" class="pa3 w-100 bg-near-white" placeholder="Password"/>
                   </div>
                   <div class="tr">
                     <button type="submit" class="pv3 ph4 bg-navy white bw0 shadow-4 ">Signup</button>
